plugin: moved definitions for MatcherCallbacks into a cpp file.
